+++ core/interaction.h
@@ -20,7 +20,7 @@ struct Interaction {
 
 	// comparison operators
 	bool operator==(const Interaction<DIM>& i) const {
-		bool distancesMatch = std::fabsf(d - i.d) < 1e-6;
+		bool distancesMatch = std::fabs(d - i.d) < 1e-6;
 		if (distanceInfo == DistanceInfo::Bounded) return distancesMatch;
 
 		return distancesMatch && (p - i.p).squaredNorm() < 1e-6;
